MiniMap
Basic functionality done, it now works well enough to release. Missing a
couple niceties though - mouseover and click for coords and info about
beacons; a way to choose where you want the map inserted on the nav
screen, cause it can be obtrusive; and doing something about the Pardus
Core, which is so large it produces a nearly unusable map.

Adding a zoom function would be nice too, and a "shortest path"... but
lets leave those for later wish issues.