Fixes for alarm not shutting up and two other issues
To trigger the alarm, we used to open a port from the msgframe to the
background page, and rely on the port being disconnected and reaped when
the frame reloads. A recent change in Chrome appears to have inhibited
this behaviour.  This change moves port management to the top frame,
which now monitors the msgframe reloading.

Additionally, a bug was fixed where the msgframe wouldn't resume
updating after a load failure (as in the server reset).

Additionally, a bug was fixed where changing the clock settings while
a Pardus tab was open would freeze the tab.