Changing browser action to page action
This avoids having to enable and disable a browser action every
time the active tab changes, like we do in Chrome. This is the
logical way anyway, because the Sweetener icon is related to the
Pardus tabs, not the entire browser. I would change this in Chrome
too, but page actions are no longer a thing in V3.